==> desembargadores/01-desembargadores.ipynb <==
Primeiro, carregamos as bibliotecas aplicáveis: o pandas para manipulação de dados, o openai para a API de geração de texto e o json para manipulação de dados JSON. O PyPDF2 é usado para ler o PDF e o dotenv para carregar as variáveis de ambiente.

import pandas as pd
import requests
from openai import OpenAI
import os
import PyPDF2
from dotenv import load_dotenv
import json
load_dotenv()

O arquivo .env contém a chave da API do OpenAI, que é carregada usando a biblioteca dotenv. Esse arquivo não está no repositório, mas você pode criar um arquivo .env com a chave da API do OpenAI, com o nome da variável `OPENAI_API_KEY`.

Abaixo, baixamos o arquivo do TJSP usando a biblioteca requests e o salvamos em um arquivo local. Em seguida, usamos o PyPDF2 para ler o arquivo e extrair o texto.

u = 'https://www.tjsp.jus.br/Download/Portal/Biblioteca/Biblioteca/Curriculum/Curriculum.pdf'
r = requests.get(u)
with open('Curriculum.pdf', 'wb') as f:
  f.write(r.content)

Esta função lê o PDF e retorna uma lista de strings, onde cada string é uma página do PDF.

def pdf_to_text_pages(pdf_file):
  # Open the PDF file in read-binary mode
  with open(pdf_file, 'rb') as file:
    # Create a PDF file reader object
    pdf_reader = PyPDF2.PdfReader(file)

    # Initialize an empty list to store the text of each page
    text_pages = []
    # Loop over the number of pages in the PDF file
    for page_num in range(len(pdf_reader.pages)):
      # Extract the text from each page and append it to the list
      page = pdf_reader.pages[page_num]
      text_pages.append(page.extract_text())

  return text_pages

# Usage
pdf_file = 'Curriculum.pdf'

text_pages = pdf_to_text_pages(pdf_file)


print(text_pages[-1])

Agora, carregamos o motor GPT-4 da OpenAI

client = OpenAI()

O prompt abaixo é a instrução inicial passada ao modelo para que ele saiba o que estamos tentando fazer. Em seguida, passaremos o texto de cada página do PDF para o modelo, que retornará os dados estruturados para cada página.

# Load the prompt from the file
with open('prompt.md', 'r') as file:
  prompt = file.read()

print(prompt)

A função abaixo é usada para enviar uma solicitação para a API do OpenAI e retornar os dados estruturados. O resultado é salvo em um arquivo JSON, para que possamos acessá-lo posteriormente caso a API fique fora do ar. Os arquivos brutos foram colocados nos [releases]() do repositório.


def pegar_texto(i_pag, text_pages):
  # if file does not exist
  file_name = f'json/desembargadores/pag_{str(i_pag).zfill(4)}.json'
  if not os.path.exists(file_name):
    completion = client.chat.completions.create(
      model="gpt-4-turbo-preview",
      messages=[
        {
          "role": "system",
          "content": prompt
        },
        {
          "role": "user",
          "content": text_pages[i_pag]
        }
      ],
      temperature=0,
      response_format={"type": "json_object"}
    )
    txt = completion.choices[0].message.content
    with open(file_name, 'a') as file:
      file.write(txt)
  else:
    with open(file_name, 'r') as file:
      txt = file.read()
  result = json.loads(txt)
  result['page'] = i_pag+1
  return result


Aqui, rodamos a função de solicitação para cada página do PDF e salvamos os resultados em arquivos JSON. O resultado é uma lista de arquivos JSON, onde cada arquivo contém os dados estruturados de uma página do PDF.

lista_desemb = []

for i in range(len(text_pages)):
  if i % 50 == 0:
    print("Pegando página: " + str(i+1) + " de " + str(len(text_pages)) + "...")
  lista_desemb.append(pegar_texto(i, text_pages))


Finalmente, carregamos os arquivos JSON e os combinamos em um único DataFrame do pandas.

df = pd.DataFrame(lista_desemb)
df

Agora, exportamos o DataFrame para um arquivo parquet, que pode ser usado para análise posterior.

df.to_parquet('parquet/desembargadores.parquet')

O próximo passo é realizar a limpeza da base, que fazemos no script `02-limpeza.r`, na linguagem R.
